% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_comparison_to_xlsx.R
\name{save_comparison_to_xlsx}
\alias{save_comparison_to_xlsx}
\title{Save summary tables made with 'compare_cols_by_group'}
\usage{
save_comparison_to_xlsx(data, file_name, sheet_name = "Sheet 1", title = "")
}
\arguments{
\item{data}{A data frame or data frame extension}

\item{file_name}{The name for the .xlsx file}

\item{sheet_name}{The XLSX worksheet to write to - can be index or name}

\item{title}{Title for the table, optional}
}
\value{
A data frame
}
\description{
Save summary tables made with 'compare_cols_by_group'
}
\examples{
data(mtcars) |>
  compare_cols_by_group(
    group_col = "cyl",
    binary_or_cat_cols = c("vs", "am", "gear", "carb"),
    continuous_cols = c("mpg", "disp", "hp", "drat", "wt", "qsec")
  ) |>
  save_comparison_to_xlsx(
    file_name = "Comparison.xlsx",
    title = "Table 1: Outcomes in Population by Group"
  )
}
