% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_cols_by_group.R
\name{summarize_cols_by_group}
\alias{summarize_cols_by_group}
\title{Make a publication-ready table}
\usage{
summarize_cols_by_group(
  data,
  group_col,
  binary_or_cat_cols = NULL,
  continuous_cols = NULL,
  remove_group_col_NA = T,
  cols_to_remove_NA = NULL,
  binary_or_cat_cols_subpop = NULL,
  binary_or_cat_cols_subpop_val = NULL,
  rounding_digits = 1
)
}
\arguments{
\item{data}{A data frame or data frame extension}

\item{group_col}{The column/variable by which the data will be grouped}

\item{binary_or_cat_cols}{Columns to be summarized as binary/categorical variables.}

\item{continuous_cols}{Columns to be summarized as continuous variables.}

\item{remove_group_col_NA}{A logical value indicating whether NA values in the 'group_col' should be dropped before the computation proceeds}

\item{cols_to_remove_NA}{Columns for which rows will be dropped if value is missing.}

\item{binary_or_cat_cols_subpop}{If specified, columns to be used for population counts for the binary_or_cat_cols.}

\item{binary_or_cat_cols_subpop_val}{Values defining the appropriate sub-populations within the 'binary_or_cat_cols_subpop' argument}

\item{rounding_digits}{Digits used for rounding table values and p-values}
}
\value{
A data frame
}
\description{
Quickly create publication-ready summary tables summarizing population sub-groups and testing for statistical differences.
}
\examples{
summary = iris |>
  summarize_cols_by_group(
    group_col = "Species",
    continuous_cols = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"),
    rounding_digits = 1
  )

summary = mtcars |>
  summarize_cols_by_group(
    group_col = "cyl",
    binary_or_cat_cols = c("vs", "am", "gear", "carb"),
    continuous_cols = c("mpg", "disp", "hp", "drat", "wt", "qsec")
  )
}
